chore(ci): nightly build
Sets up a trigger for the master build to run at midnight every night.
The deployment process is modified to check if the build was caused by
this trigger, and if so, it will upload the artifacts with a 'nightly'
tag in the file names rather than a version or commit hash.