refactor(ci): different trigger cause check
The previous check worked fine on the PR branch, but didn't seem to work
after it was merged to master. Jenkins was complaining about the
TimerTriggerCause class not existing:
groovy.lang.MissingPropertyException: No such property: TimerTriggerCause for class: groovy.lang.Binding

Trying another approach that was found on the Jenkins Google Group:
https://groups.google.com/forum/#!topic/jenkinsci-users/tKXOaXayDl0

The calls to delete the previous nightly build also need to wrapped in
the withAWS block.