feat(update): provide update mechanism
Uses the self_update crate to see if there are any updated versions of
the application available as GitHub releases.

The updating mechansim is invoked via the --update flag. If you run
safe_auth with this flag, regardless of the outcome of the update check
(i.e. if there were any updates available or if there weren't),
safe_auth will then exit. If there was an update, the next time
safe_auth runs, it'll run with the updated version.