chore(ci): new deployment process
Both the dev and release builds are now deployed. For a version change
deploy, the dev build is uploaded to S3, with the release build still
being uploaded to GitHub releases.

For a non-version commit, both the dev and release builds are uploaded
to S3.

Also fixes other little issues:
* Fixed packageBuildArtifacts method not being called correctly
* Incorrect reference to a container in the test target
* Consistently reference the build artifacts file
* Incorrect reference to Windows exe
* Incorrect reference to GH release file