fix(ci): use correct version number
We seen an issue where the release build on Linux wasn't using the
correct version number - it was using the number from the previous
release. This was because the updated code was being mounted in at the
wrong place, and the code being built was the code that was in the
container at the time it was built.